refactor: properties components
