+++ apps/journeys-admin/src/components/JourneyView/Properties/AccessControl/AccessControl.spec.tsx
@@ -1,27 +1,26 @@
 import { MockedProvider } from '@apollo/client/testing'
+import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
 import { render } from '@testing-library/react'
 import { SnackbarProvider } from 'notistack'
-import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
-import { defaultJourney } from '../data'
-import { Properties } from '.'
+import { defaultJourney } from '../../data'
+import { AccessControl } from './AccessControl'
 
 jest.mock('@mui/material/useMediaQuery', () => ({
   __esModule: true,
   default: () => true
 }))
 
-describe('JourneyView', () => {
-  it('should render journey properties', async () => {
+describe('AccessControl', () => {
+  it('should display current access', () => {
     const { getByText } = render(
       <SnackbarProvider>
-        <MockedProvider mocks={[]}>
+        <MockedProvider>
           <JourneyProvider value={{ journey: defaultJourney, admin: true }}>
-            <Properties />
+            <AccessControl />
           </JourneyProvider>
         </MockedProvider>
       </SnackbarProvider>
     )
     expect(getByText('Access Control')).toBeInTheDocument()
-    expect(getByText('Journey URL')).toBeInTheDocument()
   })
 })
