+++ apps/journeys-admin/src/components/JourneyView/Properties/AccessControl/AccessControl.tsx
@@ -1,0 +1,24 @@
+import { ReactElement } from 'react'
+import { useTranslation } from 'react-i18next'
+import { useJourney } from '@core/journeys/ui/JourneyProvider'
+import Typography from '@mui/material/Typography'
+import { AccessAvatars } from '../../../AccessAvatars'
+
+export function AccessControl(): ReactElement {
+  const { t } = useTranslation('apps-journeys-admin')
+  const { journey } = useJourney()
+
+  return (
+    <>
+      <Typography variant="subtitle2" gutterBottom>
+        {t('Access Control')}
+      </Typography>
+      <AccessAvatars
+        journeyId={journey?.id}
+        userJourneys={journey?.userJourneys ?? undefined}
+        size="medium"
+        xsMax={5}
+      />
+    </>
+  )
+}
