+++ apps/journeys-admin/src/components/JourneyView/Properties/AccessControl/index.ts
@@ -1,0 +1,1 @@
+export { AccessControl } from './AccessControl'
