+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyURL/EmbedJourneyDialog/EmbedCardPreview/EmbedCardPreview.tsx
@@ -10,8 +10,8 @@ import useMediaQuery from '@mui/material/useMediaQuery'
 import {
   ThemeName,
   ThemeMode
-} from '../../../../../../__generated__/globalTypes'
-import { FramePortal } from '../../../../FramePortal'
+} from '../../../../../../../__generated__/globalTypes'
+import { FramePortal } from '../../../../../FramePortal'
 
 const CARD_WIDTH = 340
