+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyURL/SlugDialog/SlugDialog.spec.tsx
@@ -2,7 +2,7 @@ import { MockedProvider } from '@apollo/client/testing'
 import { render, fireEvent, waitFor } from '@testing-library/react'
 import { SnackbarProvider } from 'notistack'
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
-import { defaultJourney } from '../../data'
+import { defaultJourney } from '../../../data'
 import { SlugDialog, JOURNEY_SLUG_UPDATE } from '.'
 
 const onClose = jest.fn()
