+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyURL/SlugDialog/SlugDialog.stories.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react'
 import { MockedProvider } from '@apollo/client/testing'
 import { screen, userEvent } from '@storybook/testing-library'
 import { JourneyProvider } from '@core/journeys/ui/JourneyProvider'
-import { journeysAdminConfig } from '../../../../libs/storybook'
-import { defaultJourney } from '../../data'
+import { journeysAdminConfig } from '../../../../../libs/storybook'
+import { defaultJourney } from '../../../data'
 import { SlugDialog, JOURNEY_SLUG_UPDATE } from './SlugDialog'
 
 const SlugDialogStory = {
