+++ apps/journeys-admin/src/components/JourneyView/Properties/JourneyURL/index.ts
@@ -1,0 +1,1 @@
+export { JourneyURL } from './JourneyURL'
